Add getters, extract BaseQueryTest class
